Tests/LibWeb/Text/input/canvas/fillStyle-serialization.html

<script src="../include.js"></script>
<canvas id="c" width=300 height=300></canvas>
<script>
    test(() => {
        let x = c.getContext("2d");

        function go(color) {
            x.fillStyle = color;
            println("`" + color + "` -> `" + x.fillStyle + "`");
        }

        go("green");
        go("rgba(128, 128, 128, 0.4)");
        go("rgba(128, 128, 128, 0)");
        go("rgba(128, 128, 128, 1)");
        go("rgb(128, 128, 128)");
    });
</script>
