Tests/LibWeb/Text/expected/canvas/fillStyle-serialization.txt

`green` -> `#008000`
`rgba(128, 128, 128, 0.4)` -> `rgba(128, 128, 128, 0.4)`
`rgba(128, 128, 128, 0)` -> `rgba(128, 128, 128, 0)`
`rgba(128, 128, 128, 1)` -> `#808080`
`rgb(128, 128, 128)` -> `#808080`
